계약 1차 구현

+++ khedu/src/main/java/com/kh/khedu/vo/payroll/request/AttendanceAbsentRequestVO.java
@@ -10,7 +10,7 @@ import lombok.Data;
 public class AttendanceAbsentRequestVO {
 
     @Positive
-    private long employeeNo;
+    private int employeeNo;
 
     @NotNull
     private Timestamp workDate;
